+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 3.7.4 (2014-xx-xx)
 --
+Improvements:
+
+* Profiles: mass import for profiles using a .csv added.
+
 Bugfixes:
 
 * Core: event subscriptions did not get fired in the frontend.